Add settings cog
The settings cog manages settings for the bot overall and per guild. It allows
guild owners to manage settings for their specific server and bot owners to
manage settings for the entire fleet of guilds using their bot. It's a
dependency for many other Unbuffed Bot Cogs.

Settings can be restricted globally (making it impossible for guild owners to
override them) and on a per guild level (making it impossible for the speciifc
 guild to change their current value).