Add Steam cog
The Steam cog allows interfacing with the steam store. It currently can search
the steam store for a set of results using the `!steam search` command and
return details for a specific given game using the `!steam game` command.

It utilizes the sqlite Cog to create a cache table to cache game details up to a
day.