Adds LetsTry cog and supporting libraries
This is a big one. And yet it's also not done done. It is functional though, so
I decided to pack it up into a commit and deploy.
Additional changes and improvements, including more tests, better command help
and additional convenience functions are going to be added in the coming days
and weeks in additional commits.